# [libc++] Optimize copy construction and assignment of __tree